Split the stores
