+++ src/components/ManagePoints.vue
@@ -1,10 +1,11 @@
 <script setup lang="ts">
 import { type Ref, ref, computed } from 'vue';
 import type { Player } from '@/types';
-import { usePlayersStore } from '@/stores/players';
+import { useGameStore, usePlayersStore } from '@/stores';
 
-const store = usePlayersStore();
-const playersRoundData = computed(() => store.players);
+const gameStore = useGameStore();
+const playersStore = usePlayersStore();
+const playersRoundData = computed(() => playersStore.players);
 const enableCounting: Ref<boolean> = ref(false);
 // TO TEST
 // const enableCounting: Ref<boolean> = ref(true);
@@ -33,14 +34,14 @@ function updatePlayersPoints() {
 function closeRound() {
   if (confirm("Es-tu sûr d'avoir fini le tour x ?") === true) {
     updatePlayersPoints();
-    store.updatePlayers(playersRoundData.value);
+    playersStore.updatePlayers(playersRoundData.value);
     enableCounting.value = false;
   }
 }
 </script>
 
 <template>
-  <section v-if="store.gameStarted" class="dominos">
+  <section v-if="gameStore.gameStarted" class="dominos">
     <h1>Domino: 12</h1>
     <button
       v-if="!enableCounting"
