+++ src/components/ResetButtons.vue
@@ -1,12 +1,13 @@
 <script setup lang="ts">
-import { usePlayersStore } from '@/stores/players';
+import { useGameStore, usePlayersStore } from '@/stores';
 
-const store = usePlayersStore();
+const gameStore = useGameStore();
+const playersStore = usePlayersStore();
 </script>
 
 <template>
-  <button type="button" @click="store.resetGame">Recommencer la partie</button>
-  <button type="button" @click="store.resetPlayers">Changer de joueurs</button>
+  <button type="button" @click="gameStore.resetGame">Recommencer la partie</button>
+  <button type="button" @click="playersStore.resetPlayers">Changer de joueurs</button>
   <button type="button" @click="alert('TODO: Historique')">Historique</button>
   <button type="button" @click="alert('TODO: règles')">Règles</button>
 </template>
