+++ src/stores/index.ts
@@ -1,0 +1,2 @@
+export * from './GameStore';
+export * from './PlayersStore';
